requirements.txt
streamlit
opencv-python
pytesseract
pillow
google-generativeai
python-dotenv
reportlab
python-docx
pandas
easyocr